requirements.txt
google-api-python-client==2.169.0
pymongo==4.6.1
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.1
pytz==2024.1 
aiohttp==3.9.3
paramiko==3.4.0
pydantic==2.6.1
pika==1.3.2
kafka-python==2.0.2
python-dateutil==2.9.0
